[#354] Optimize CoroutineTestRule & TestDispatchersProvider usage
