--- sample-compose/app/src/androidTest/java/co/nimblehq/sample/compose/test/CoroutineTestRule.kt
@@ -1,32 +1,0 @@
-package co.nimblehq.sample.compose.test
-
-import co.nimblehq.sample.compose.util.DispatchersProvider
-import kotlinx.coroutines.*
-import kotlinx.coroutines.test.*
-import org.junit.rules.TestWatcher
-import org.junit.runner.Description
-
-@OptIn(ExperimentalCoroutinesApi::class)
-class CoroutineTestRule(
-    private val testDispatcher: TestDispatcher = UnconfinedTestDispatcher()
-) : TestWatcher() {
-
-    override fun starting(description: Description?) {
-        super.starting(description)
-        Dispatchers.setMain(testDispatcher)
-    }
-
-    override fun finished(description: Description?) {
-        super.finished(description)
-        Dispatchers.resetMain()
-    }
-
-    val testDispatcherProvider = object : DispatchersProvider {
-        override val io: CoroutineDispatcher
-            get() = testDispatcher
-        override val main: CoroutineDispatcher
-            get() = testDispatcher
-        override val default: CoroutineDispatcher
-            get() = testDispatcher
-    }
-}
