+++ sample-compose/app/src/test/java/co/nimblehq/sample/compose/test/CoroutineTestRule.kt
@@ -11,22 +11,23 @@ class CoroutineTestRule(
     private val testDispatcher: TestDispatcher = UnconfinedTestDispatcher()
 ) : TestWatcher() {
 
+    val testDispatcherProvider = object : DispatchersProvider {
+
+        override val io: CoroutineDispatcher
+            get() = testDispatcher
+
+        override val main: CoroutineDispatcher
+            get() = testDispatcher
+
+        override val default: CoroutineDispatcher
+            get() = testDispatcher
+    }
+
     override fun starting(description: Description?) {
-        super.starting(description)
         Dispatchers.setMain(testDispatcher)
     }
 
     override fun finished(description: Description?) {
-        super.finished(description)
         Dispatchers.resetMain()
-    }
-
-    val testDispatcherProvider = object : DispatchersProvider {
-        override val io: CoroutineDispatcher
-            get() = testDispatcher
-        override val main: CoroutineDispatcher
-            get() = testDispatcher
-        override val default: CoroutineDispatcher
-            get() = testDispatcher
     }
 }
